requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
openai>=1.12.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
python-socketio>=5.11.1
aiohttp>=3.9.3 